+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Current Master
 
+- Fix bug where the "Reset account" feature would not clear the network cache.
+- Increase maximum gas limit, to allow very gas heavy transactions, since block gas limits have been stable.
+- Allow user to select language on settings page
+
 ## 4.4.0 Mon Mar 26 2018
 
 - Internationalization: Taiwanese, Thai, Slovenian